refactor: modularize daggerheart app layout
